Apply suggestions from CR
